Refactor drone service packaging and add a new intermediate interface DroneDataEditable.kt to mock it easily
