+++ app/src/main/java/ch/epfl/sdp/drone3d/ui/drone/ConnectedDroneActivity.kt
@@ -11,7 +11,7 @@ import android.view.View
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import ch.epfl.sdp.drone3d.R
-import ch.epfl.sdp.drone3d.drone.DroneService
+import ch.epfl.sdp.drone3d.drone.api.DroneService
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
